fields for gramage

+++ pojo-ops/src/main/java/com/stanzaliving/core/food/dto/response/DishMasterRatingResponseDto.java
@@ -27,7 +27,7 @@ public class DishMasterRatingResponseDto extends DishMasterSearchResponseDto {
 
 	private static final long serialVersionUID = -77978649022785L;
 
-	private Double itemGrammage;
+	private Integer itemGrammage;
 
 	private Integer itemQty;
